+++ solidity/analysis/contracts/Loop.sol
@@ -2,19 +2,19 @@
 pragma solidity >=0.8.0 <0.9.0;
 
 contract Loop {
-    uint num = 0; // Storage
-    
-    function expensiveLoop(uint x) public {
-        for(uint i = 0; i < x; i++) {
-            num += 1;
-        }
-    }
-    
-    function lessExpensiveLoop(uint x) public {
-        uint temp = num;
-        for(uint i = 0; i < x; i++) {
-            temp += 1;
-        }
-        num = temp;
-    }
+	uint256 num = 0; // Storage
+
+	function expensiveLoop(uint256 x) public {
+		for (uint256 i = 0; i < x; i++) {
+			num += 1;
+		}
+	}
+
+	function lessExpensiveLoop(uint256 x) public {
+		uint256 temp = num;
+		for (uint256 i = 0; i < x; i++) {
+			temp += 1;
+		}
+		num = temp;
+	}
 }
