+++ solidity/analysis/contracts/MappingsAnnidati.sol
@@ -2,13 +2,13 @@
 pragma solidity >=0.8.0 <0.9.0;
 
 contract MappingsAnnidati {
-    struct S { 
-        uint128 a; 
-        uint128 b; 
-        uint256 c;
-    }
-    
-    uint x;
-    
-    mapping(uint => mapping(uint => S)) data;
+	struct S {
+		uint128 a;
+		uint128 b;
+		uint256 c;
+	}
+
+	uint256 x;
+
+	mapping(uint256 => mapping(uint256 => S)) data;
 }
